only match each event.type on the first encounter
Introduced 'typeFixer' functions with associated regexes for event
types. Each time we encounter a new event type, look up the fixer
function for that type (there is a default) by testing against the
regexes. Store that fixer function in a cache so next time we have
to handle that event type we just fetch it from the cache and
execute it. Fixer functions are also responsible for dictating the
properties to be copied from the original event to the synthetic.